chore: put all task entrypoints into plugin toolkitrcs
