--- plugins/babel/src/index.ts
@@ -1,4 +1,0 @@
-import BabelDevelopment from './tasks/development'
-import BabelProduction from './tasks/production'
-
-export const tasks = { BabelDevelopment, BabelProduction }
