--- plugins/circleci/.toolkitrc.yml
@@ -1,3 +1,0 @@
-installs:
-  - 'build:ci'
-  - 'test:ci'
