--- plugins/cypress/src/index.ts
@@ -1,3 +1,0 @@
-import { CypressLocal, CypressCi } from './tasks/cypress'
-
-export const tasks = { CypressLocal, CypressCi }
