+++ plugins/cypress/src/tasks/ci.ts
@@ -3,29 +3,8 @@ import { hookFork, waitOnExit } from '@dotcom-tool-kit/logger'
 import { readState } from '@dotcom-tool-kit/state'
 import { Task } from '@dotcom-tool-kit/types'
 import { CypressSchema } from '@dotcom-tool-kit/types/lib/schema/plugins/cypress'
-import { VaultEnvVars } from '@dotcom-tool-kit/vault'
 
-export class CypressLocal extends Task<typeof CypressSchema> {
-  async run(): Promise<void> {
-    const cypressEnv: Record<string, string> = {}
-    if (this.options.localUrl) {
-      cypressEnv.CYPRESS_BASE_URL = this.options.localUrl
-    }
-
-    const vault = new VaultEnvVars(this.logger, {
-      environment: 'development'
-    })
-    const vaultEnv = await vault.get()
-
-    const env = { ...process.env, ...cypressEnv, ...vaultEnv }
-    this.logger.info('running cypress' + (env.CYPRESS_BASE_URL ? ` against ${env.CYPRESS_BASE_URL}` : ''))
-    const testProcess = spawn('cypress', ['run'], { env })
-    hookFork(this.logger, 'cypress', testProcess)
-    return waitOnExit('cypress', testProcess)
-  }
-}
-
-export class CypressCi extends Task<typeof CypressSchema> {
+export default class CypressCi extends Task<typeof CypressSchema> {
   async run(): Promise<void> {
     const reviewState = readState('review')
     const cypressEnv: Record<string, string> = {}
