--- plugins/eslint/src/index.ts
@@ -1,3 +1,0 @@
-import Eslint from './tasks/eslint'
-
-export const tasks = { Eslint }
