--- plugins/jest/src/index.ts
@@ -1,7 +1,0 @@
-import JestCI from './tasks/ci'
-import JestLocal from './tasks/local'
-
-export const tasks = {
-  JestCI,
-  JestLocal
-}
