--- plugins/lint-staged/src/hook.ts
@@ -1,6 +1,0 @@
-import { PackageJsonHelper } from '@dotcom-tool-kit/package-json-hook'
-
-export abstract class LintStagedHook extends PackageJsonHelper {
-  field = 'lint-staged'
-  trailingString = '--'
-}
