--- plugins/lint-staged/src/index.ts
@@ -1,4 +1,0 @@
-import LintStaged from './tasks/lint-staged'
-
-export const tasks = { LintStaged }
-export { LintStagedHook } from './hook'
