--- plugins/mocha/src/index.ts
@@ -1,3 +1,0 @@
-import Mocha from './tasks/mocha'
-
-export const tasks = { Mocha }
