+++ plugins/n-test/.toolkitrc.yml
@@ -1,5 +1,5 @@
 tasks:
-  - NTest
+  NTest: './lib/tasks/n-test'
 
 commands:
   'test:review': NTest
