--- plugins/n-test/src/index.ts
@@ -1,3 +1,0 @@
-import NTest from './tasks/n-test'
-
-export const tasks = { NTest }
