--- plugins/next-router/src/index.ts
@@ -1,3 +1,0 @@
-import NextRouter from './tasks/next-router'
-
-export const tasks = { NextRouter }
