--- plugins/node/src/index.ts
@@ -1,3 +1,0 @@
-import Node from './tasks/node'
-
-export const tasks = { Node }
