+++ plugins/nodemon/.toolkitrc.yml
@@ -2,7 +2,7 @@ plugins:
   - '@dotcom-tool-kit/doppler'
 
 tasks:
-  - Nodemon
+  Nodemon: './lib/tasks/nodemon'
 
 commands:
   'run:local': Nodemon
