--- plugins/nodemon/src/index.ts
@@ -1,3 +1,0 @@
-import Nodemon from './tasks/nodemon'
-
-export const tasks = { Nodemon }
