+++ plugins/npm/.toolkitrc.yml
@@ -1,3 +1,3 @@
 tasks:
-  - NpmPrune
-  - NpmPublish
+  NpmPrune: './lib/tasks/prune'
+  NpmPublish: './lib/tasks/publish'
