+++ plugins/npm/src/index.ts
@@ -1,10 +1,1 @@
-import { PackageJsonScriptHook } from '@dotcom-tool-kit/package-json-hook'
-import NpmPrune from './tasks/npm-prune'
-import NpmPublish from './tasks/npm-publish'
-
-export { PackageJsonScriptHook as PackageJsonScriptHook }
-
-export const tasks = {
-  NpmPrune,
-  NpmPublish
-}
+export { PackageJsonScriptHook } from '@dotcom-tool-kit/package-json-hook'
