+++ plugins/npm/src/tasks/prune.ts
