+++ plugins/pa11y/.toolkitrc.yml
@@ -2,7 +2,7 @@ plugins:
   - '@dotcom-tool-kit/doppler'
 
 tasks:
-  - Pa11y
+  Pa11y: './lib/tasks/pa11y'
 
 commands:
   'test:review': Pa11y
