--- plugins/prettier/src/index.ts
@@ -1,3 +1,0 @@
-import Prettier from './tasks/prettier'
-
-export const tasks = { Prettier }
