--- plugins/serverless/src/index.ts
@@ -1,5 +1,0 @@
-import ServerlessDeploy from './tasks/deploy'
-import ServerlessProvision from './tasks/provision'
-import ServerlessRun from './tasks/run'
-
-export const tasks = { ServerlessRun, ServerlessDeploy, ServerlessProvision }
