+++ plugins/typescript/.toolkitrc.yml
@@ -1,7 +1,7 @@
 tasks:
-  - TypeScriptBuild
-  - TypeScriptWatch
-  - TypeScriptTest
+  TypeScriptBuild: './lib/tasks/build'
+  TypeScriptWatch: './lib/tasks/watch'
+  TypeScriptTest: './lib/tasks/test'
 
 commands:
   'build:local': TypeScriptBuild
