--- plugins/typescript/src/index.ts
@@ -1,3 +1,0 @@
-import { TypeScriptBuild, TypeScriptWatch, TypeScriptTest } from './tasks/typescript'
-
-export const tasks = { TypeScriptBuild, TypeScriptWatch, TypeScriptTest }
