+++ plugins/typescript/src/tasks/build.ts
@@ -1,0 +1,7 @@
+import TypeScriptTask from './typescript'
+
+export default class TypeScriptBuild extends TypeScriptTask {
+  static description = 'compile TypeScript to JavaScript'
+
+  taskArgs = []
+}
