+++ plugins/typescript/src/tasks/test.ts
@@ -1,0 +1,7 @@
+import TypeScriptTask from './typescript'
+
+export default class TypeScriptTest extends TypeScriptTask {
+  static description = 'type check TypeScript code'
+
+  taskArgs = ['--noEmit']
+}
