+++ plugins/upload-assets-to-s3/.toolkitrc.yml
@@ -1,5 +1,5 @@
 tasks:
-  - UploadAssetsToS3
+  UploadAssetsToS3: './lib/tasks/upload-assets-to-s3'
 
 commands:
   'release:remote': UploadAssetsToS3
