--- plugins/upload-assets-to-s3/src/index.ts
@@ -1,3 +1,0 @@
-import UploadAssetsToS3 from './tasks/upload-assets-to-s3'
-
-export const tasks = { UploadAssetsToS3 }
