--- plugins/webpack/src/index.ts
@@ -1,5 +1,0 @@
-import WebpackDevelopment from './tasks/development'
-import WebpackProduction from './tasks/production'
-import WebpackWatch from './tasks/watch'
-
-export const tasks = { WebpackDevelopment, WebpackProduction, WebpackWatch }
